Fix critical availability bug, harden security, and redesign demo UI
Backend audit and hardening:
- Fix a critical bug where every real HTTP request failed with 500
  ("Unable To Extract Key!") because tower_governor's rate limiter
  requires ConnectInfo that main.rs never wired up via
  into_make_service_with_connect_info. Integration tests missed this
  because they inject ConnectInfo manually.
- Fix an IDOR: authenticated devices could read any other device's
  sync documents and history by guessing/enumerating a UUID, since
  get_document_handler/get_document_history_handler never checked
  ownership. Now scoped to the requesting device, with a regression
  test.
- Key the audit-log hash chain with HMAC-SHA256 (domain-separated from
  the session-token key) instead of a plain unkeyed SHA-256 digest, so
  forging a consistent chain requires the server secret, not just
  database write access.
- Lock down CORS to specific methods/headers outside development
  (previously wide open in every environment); a native iOS client is
  unaffected since it never sends Origin.
- Correct a misleading "https://" log/README claim — the engine speaks
  plain HTTP and needs a TLS-terminating proxy in front in production.
- Clean up all clippy::pedantic findings (uninlined format args,
  lossy casts, redundant continue, missing must_use, doc formatting).

Demo UI redesign (demo/index.html):
- Replaced the gradient/emoji/glow "AI slop" look with a flat,
  hairline-bordered, monospace-driven instrument aesthetic (validated
  via a Figma reference mockup and the dataviz color-formula skill).
- Same functionality (dashboard, AI chat, sync, live API console)
  end-to-end verified against the running backend in a real browser.
- Fixed a layout regression caught during that verification: the chat
  screen had no bottom padding to clear the fixed tab bar, so the tab
  bar intercepted clicks on the send button.

+++ src/crypto/assertion.rs
@@ -11,8 +11,8 @@ pub struct VerifiedAssertion {
 /// Steps:
 /// 1. Decode CBOR assertion object to extract authenticatorData and signature
 /// 2. Parse authenticatorData to verify rpIdHash and flags
-/// 3. Verify that the assertion's signCount is strictly greater than the stored_counter
-/// 4. Reconstruct clientDataHash = SHA-256(challenge || request_body_hash)
+/// 3. Verify that the assertion's signCount is strictly greater than the `stored_counter`
+/// 4. Reconstruct clientDataHash = SHA-256(challenge || `request_body_hash`)
 /// 5. Verify ECDSA signature over (authenticatorData || clientDataHash)
 pub fn verify_assertion(
     config: &AppConfig,
@@ -85,12 +85,12 @@ pub fn verify_assertion(
     let rp_id_hash = &authenticator_data_bytes[0..32];
     let flags = authenticator_data_bytes[32];
 
-    let sign_count = u32::from_be_bytes([
+    let sign_count = i64::from(u32::from_be_bytes([
         authenticator_data_bytes[33],
         authenticator_data_bytes[34],
         authenticator_data_bytes[35],
         authenticator_data_bytes[36],
-    ]) as i64;
+    ]));
 
     // Verify rpIdHash matches SHA-256 of App ID
     let expected_app_id = config.auth.app_id();
